split form and add second step
