+++ frontend/packages/client/src/components/ProposalCreate/StepTwo/ImageChoices.js
